requirements.txt
-i https://pypi.org/simple
blinker==1.4
certifi==2023.11.17 ; python_version >= '3.6'
click==7.1.2
flask==1.1.2
flask-debugtoolbar==0.11.0
itsdangerous==1.1.0
jinja2==2.11.3
markupsafe==1.1.1
pymongo==3.11.3
python-dotenv==0.16.0
sentry-sdk[flask]==1.9.0
urllib3==1.26.6
werkzeug==1.0.1
